// src/app/main/personas/models/persona.ts
export class CompletarPerfil {
    identificacion: string;
    nombres: string;
    apellidos: string;
    genero: string;
    fechaNacimiento: string;
    edad: number;
    whatsapp: string;
    user_id: string;
}
export class InformacionBasica {
    genero: string;
    fechaNacimiento: string;
    edad: number;
    ciudad: string;
    emailAdicional: string;
    whatsapp: string;
    facebook: string;
    instagram: string;
    twitter: string;
    tiktok: string;
    youtube: string;
    user_id: string;
}